Add option to make browsers work with comma-separated string.
I want to be able to specify browsers as part of a Grunt template. I want to
easily specify what browsers to run as part of a task call like

    grunt karma:unitTest:appName:Chrome,Firefox,IE

or even use a single browser like

    grunt karma:unitTest:appName:Chrome

and this is quite easy through Grunt templates by stating

 browsers: '<%= grunt.task.current.args[1] %>'

But we need to make browsers work with a string input, and not just an array
input. This change will split the browsers definition in the gruntfile on
commas if it is a string. It will not change anything if an array is used.

Since I have made this feature more complicated, I have added a new debug
log message to tell the user what browsers we are using to help resolve any
issues.